[ALPS03682533] disp: fix compare condition for unsigned variable
do cast to signed vailable before compare condiiton

MTK-Commit-Id: bc2faf216ca40925149bb919159d3544b7173e99

CR-Id: ALPS03682533
Feature: Others

+++ drivers/misc/mediatek/video/mt6765/videox/disp_lowpower.c
@@ -896,11 +896,11 @@ static int _primary_path_idlemgr_monitor_thread(void *data)
 			MMPROFILE_FLAG_PULSE, idle_check_interval, interval);
 
 		/* error handling */
-		interval = interval > 1000 ? 1000 : interval;
+		interval = (long long)interval > 1000 ? 1000 : interval;
 		/* when starting up before the first time kick */
 		if (idlemgr_pgc->idlemgr_last_kick_time == 0)
 			msleep_interruptible(idle_check_interval);
-		else if (interval > 0)
+		else if ((long long)interval > 0)
 			msleep_interruptible(interval);
 
 		primary_display_manual_lock();
